Cleanup pt2
- Reverted GetPlayerAvatarGenderByGraphicsId return type
- Removed IsAccurateGBA
- Removed TIME_OF_DAY constants

+++ include/field_player_avatar.h
@@ -4,7 +4,7 @@
 void PlayerStep(u8 direction, u16 newKeys, u16 heldKeys);
 void ClearPlayerAvatarInfo(void);
 void SetPlayerAvatarExtraStateTransition(u16, u8);
-u16 GetPlayerAvatarGenderByGraphicsId(u16);
+u8 GetPlayerAvatarGenderByGraphicsId(u16);
 bool8 TestPlayerAvatarFlags(u8);
 u8 GetPlayerAvatarSpriteId(void);
 void PlayerGetDestCoords(s16 *, s16 *);
